Add support for `request_id` in `GRPC.Server.Stream` (#198)
* Add support for `request_id` in `GRPC.Server.Stream`

* Apply suggestions from code review

* chore: format

* test: add tests

* chore: format

* chore: revert

* fix: revert some more
