+++ lib/grpc/server/stream.ex
@@ -25,6 +25,7 @@ defmodule GRPC.Server.Stream do
           endpoint: atom,
           rpc: tuple,
           request_mod: atom,
+          request_id: String.t() | nil,
           response_mod: atom,
           codec: atom,
           payload: any,
@@ -43,6 +44,7 @@ defmodule GRPC.Server.Stream do
             endpoint: nil,
             rpc: nil,
             request_mod: nil,
+            request_id: nil,
             response_mod: nil,
             codec: GRPC.Codec.Proto,
             payload: nil,
